fix: production-readiness audit follow-up + CI/CD test fixes
- Fix N+1 queries: JOIN FETCH added to DailyLog, LeaveRequest,
  AuditLog, Notification, Employee, Task, CashCollection,
  Holiday, Payroll repositories
- Add SLF4J loggers to all services, replace System.err.println
  with log.warn
- Fix JWT secret length logging level to DEBUG
- Fix DailyLogService local variable 'log' shadowing logger field
- Fix RecycleBinController ApiResponse type mismatch
- Fix CI/CD: add @ActiveProfiles('test') to all 4 controller tests
- Expand application-test.yml with all required env var defaults
- Fix test BigDecimal type mismatches in DailyLogServiceTest,
  SalaryAdvanceServiceTest
- Fix AuthController register endpoint to return 201
- Fix AuthServiceTest resetPassword to hash token correctly
- Fix AuthServiceTest loginWithGoogle to add email to allow list
- Fix AuthControllerTest register to include all required fields

+++ allinone-backend/src/test/java/com/workflow/service/DailyLogServiceTest.java
@@ -11,6 +11,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.ActiveProfiles;
 
 import java.math.BigDecimal;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
@@ -92,7 +93,7 @@ class DailyLogServiceTest {
         DailyLogRequest request = createRequest(employee.getId());
         request.setStartTime("09:00");
         request.setEndTime("17:00");
-        request.setHoursWorked(8.0);
+        request.setHoursWorked(BigDecimal.valueOf(8.0));
         DailyLog result = dailyLogService.createDailyLog(request, employee.getId());
         assertEquals(LocalTime.of(9, 0), result.getStartTime());
         assertEquals(LocalTime.of(17, 0), result.getEndTime());
